move replication jobs into Replication namespace
