--- app/jobs/delivery_dispatcher_job.rb
@@ -1,63 +1,0 @@
-# frozen_string_literal: true
-
-# Precondition(s):
-#  All needed PreservedObject and ZippedMoabVersion rows are already made.
-# @see PreservedObject#create_zipped_moab_versions!
-#
-# Responsibilities:
-# Record zip part metadata info in DB.
-# Split message out to all necessary zip endpoints.
-# For example:
-#   Endpoint1Delivery.perform_later(druid, version, part_s3_key)
-#   Endpoint2Delivery.perform_later(druid, version, part_s3_key)
-#   ...
-#
-# Do not assume we can just get metadata from (the DruidVersionZip) zip.
-# Jobs are not run at the same time or on the same system, so the info may not match.
-# Therefore, we receive the info passed by the process that was there when the file was created.
-class DeliveryDispatcherJob < ZipPartJobBase
-  queue_as :zips_made
-
-  before_enqueue do |job|
-    job.zip_info_check!(job.arguments.fourth)
-  end
-
-  # @param [String] druid
-  # @param [Integer] version
-  # @param [String] part_s3_key, e.g. 'ab/123/cd/4567/ab123cd4567.v0001.z03'
-  # @param [Hash<Symbol => String>] metadata Zip info
-  # @option metadata [Integer] :parts_count
-  # @option metadata [Integer] :size
-  # @option metadata [String] :checksum_md5
-  # @option metadata [String] :suffix
-  # @option metadata [String] :zip_cmd
-  # @option metadata [String] :zip_version
-  def perform(druid, version, part_s3_key, metadata)
-    zmvs.each do |zmv|
-      find_or_create_unreplicated_part(zmv, part_s3_key, metadata)
-    end
-    deliverers.each { |worker| worker.perform_later(druid, version, part_s3_key, metadata) }
-  end
-
-  private
-
-  # @return [ActiveRecord::Relation] effectively an Array of ZippedMoabVersion objects
-  def zmvs
-    @zmvs ||= ZippedMoabVersion.by_druid(zip.druid.id).where(version: zip.version)
-  end
-
-  def find_or_create_unreplicated_part(zmv, part_s3_key, metadata)
-    zmv.zip_parts.create_with(
-      create_info: metadata.slice(:zip_cmd, :zip_version).to_s,
-      md5: metadata[:checksum_md5],
-      parts_count: metadata[:parts_count],
-      size: metadata[:size],
-      suffix: File.extname(part_s3_key)
-    ).create_or_find_by(suffix: File.extname(part_s3_key), &:unreplicated!)
-  end
-
-  # @return [Array<Class>] target delivery worker classes
-  def deliverers
-    zmvs.map { |zmv| zmv.zip_endpoint.delivery_class.constantize }.uniq
-  end
-end
