+++ config/settings/development.yml
@@ -18,7 +18,7 @@ zip_endpoints:
     region: 'localhost'
     endpoint_node: 'localhost'
     storage_location: 'bucket_name'
-    delivery_class: 'S3WestDeliveryJob'
+    delivery_class: 'Replication::S3WestDeliveryJob'
     audit_class: 'Audit::ReplicationToAws'
     access_key_id: 'foo'
     secret_access_key: 'bar'
