+++ config/settings/test.yml
@@ -19,7 +19,7 @@ zip_endpoints:
     region: 'us-west-2'
     endpoint_node: 'us-west-2'
     storage_location: 'sul-sdr-aws-us-west-2-test'
-    delivery_class: 'S3WestDeliveryJob'
+    delivery_class: 'Replication::S3WestDeliveryJob'
     audit_class: 'Audit::ReplicationToAws'
     access_key_id: 'overridden-by-env-var-in-ci'
     secret_access_key: 'overridden-by-env-var-in-ci'
@@ -27,7 +27,7 @@ zip_endpoints:
     region: 'us-south'
     endpoint_node: 'https://s3.us-south.cloud-object-storage.appdomain.cloud'
     storage_location: 'sul-sdr-ibm-us-south-1-test'
-    delivery_class: 'IbmSouthDeliveryJob'
+    delivery_class: 'Replication::IbmSouthDeliveryJob'
     audit_class: 'Audit::ReplicationToIbm'
     access_key_id: 'overridden-by-env-var-in-ci'
     secret_access_key: 'overridden-by-env-var-in-ci'
