+++ spec/jobs/replication/ibm_south_delivery_job_spec.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+describe Replication::IbmSouthDeliveryJob do
+  it 'descends from Replication::AbstractDeliveryJob' do
+    expect(described_class.new).to be_an(Replication::AbstractDeliveryJob)
+  end
+
+  it 'uses its own queue' do
+    expect(described_class.new.queue_name).to eq 'ibm_us_south_delivery'
+  end
+end
