+++ spec/jobs/replication/s3_east_delivery_job_spec.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+describe Replication::S3EastDeliveryJob do
+  it 'descends from Replication::AbstractDeliveryJob' do
+    expect(described_class.new).to be_an(Replication::AbstractDeliveryJob)
+  end
+
+  it 'uses its own queue' do
+    expect(described_class.new.queue_name).to eq 's3_us_east_1_delivery'
+  end
+end
